More cleanup

+++ Sources/libcoverage/coverage.c
@@ -63,7 +63,7 @@ int cov_initialize(struct cov_context* context)
 
 void cov_finish_initialization(struct cov_context* context, int should_track_edges)
 {
-    uint64_t num_edges = context->shmem->num_edges;
+    uint32_t num_edges = context->shmem->num_edges;
     if (num_edges == 0) {
         fprintf(stderr, "[LibCoverage] Coverage bitmap size could not be determined, is the engine instrumentation working properly?\n");
         exit(-1);
@@ -78,7 +78,7 @@ void cov_finish_initialization(struct cov_context* context, int should_track_edg
         exit(-1);           // TODO
     }
 
-    uint64_t bitmap_size = (num_edges + 7) / 8; //Num edges in bytes
+    uint32_t bitmap_size = (num_edges + 7) / 8; //Num edges in bytes
 
     context->num_edges = num_edges;
     context->bitmap_size = bitmap_size;
@@ -92,9 +92,9 @@ void cov_finish_initialization(struct cov_context* context, int should_track_edg
     memset(context->crash_bits, 0xff, bitmap_size);
 
     if (should_track_edges) {
-        context->edge_count = malloc(sizeof(uint64_t) * num_edges);
+        context->edge_count = malloc(sizeof(uint32_t) * num_edges);
         assert(context->edge_count != NULL);
-        memset(context->edge_count, 0, sizeof(uint64_t) * num_edges);
+        memset(context->edge_count, 0, sizeof(uint32_t) * num_edges);
     } else {
         context->edge_count = NULL;
     }
@@ -172,7 +172,7 @@ int cov_evaluate_crash(struct cov_context* context)
     return num_new_edges > 0;
 }
 
-int cov_compare_equal(struct cov_context* context, uint64_t* edges, uint64_t num_edges)
+int cov_compare_equal(struct cov_context* context, uint32_t* edges, uint64_t num_edges)
 {
     for (int i = 0; i < num_edges; i++) {
         int idx = edges[i];
